Refactor: rename uninsured files and update paths
